refactor: 게임 재시작 시 이름 입력칸에 포커싱, remove함수 view에 모아둠
