+++ src/controller/play.ts
@@ -3,6 +3,7 @@ import { Car, Game } from '../model/index.js';
 import { renderArrowDiv, renderResultSection } from '../view/renderer.js';
 import { setRestartButtonEvent } from './event.js';
 import { $$, makeDelay } from '../utils.js';
+import { removeSpinners } from '../view/remover.js';
 
 const createCarsObject = (carNameArray: Array<string>) => {
   let cars: Array<Car> = [];
@@ -29,14 +30,13 @@ const playOnce = (racingGame: Game) => {
 
 const startGame = async (carNameArray: Array<string>, tryCount: number) => {
   const racingGame: Game = new Game(createCarsObject(carNameArray));
+  let racingWinners: Array<string> = [];
 
   for (let index = 0; index < tryCount; index += 1) {
     await makeDelay(1000).then(() => playOnce(racingGame));
   } // 시도 횟수만큼 플레이
-  $$('div.d-flex.justify-center.mt-3').forEach((element) => {
-    element.remove();
-  }); // 게임 끝나면 spinner 제거
-  const racingWinners: Array<string> = racingGame.getWinners();
+  removeSpinners();
+  racingWinners = racingGame.getWinners();
   renderResultSection(racingWinners.join(', ').toLowerCase());
   await makeDelay(2000).then(() => alert(racingWinners.join(', ') + ALERT.CONGRATULATION));
   setRestartButtonEvent();
